Remove functions.d.ts
