+++ VocaDbWeb/Scripts/Tests/setup.js
@@ -1,3 +1,2 @@
 vdb = {};
-vdb.functions = vdb.functions || {};
 vdb.values = vdb.values || {};
