--- VocaDbWeb/Scripts/typings/vdb/functions.d.ts
@@ -1,5 +1,0 @@
-﻿declare module vdb.functions {
-	export function boldCaseInsensitive(text: string, term: string): string;
-
-	export function disableTabReload(tab: any);
-}
