+++ VocaDbWeb/Views/Shared/GlobalSearchBox.cshtml
@@ -52,7 +52,7 @@
 
 		<ul class="dropdown-menu" role="menu" aria-labelledby="dropdownMenu">
 			@if (!Login.Manager.IsLoggedIn) {
-				<li><a href="@Url.Action("Login", "User")" onclick=" vdb.functions.showLoginPopup(); return false; ">@ViewRes._LayoutStrings.LogIn</a></li>
+				<li><a href="@Url.Action("Login", "User")" onclick=" app.functions.showLoginPopup(); return false; ">@ViewRes._LayoutStrings.LogIn</a></li>
 				<li><a asp-action="Create" asp-controller="User">@(ViewRes._LayoutStrings.Register)</a></li>
 			} else {
 				<li><a asp-action="Profile" asp-controller="User" asp-route-id="@(Login.Manager.Name)">@(ViewRes.User.MySettingsStrings.Profile)</a></li>
